Use a single vararg instead of command + args...
Having a separate `command` arg separate from `args` seemed useful, but
actually complicated the handling of the parsed options. When using the
`TrailingVarArg` option with `required(true)`, it's guaranteed that
there will be at least one arg present, so we can safely force-unwrap
the command name and then iterate over the remaining args if present.

This also improves the help text, where the command/args distinction was
actually somewhat confusing.